Change location of image generation
